pyopengl2-19 mapping cursor pos into view coordinate system
resolved #19
