+++ .idea/shelf/Uncommitted_changes_before_Checkout_at_2020-08-06__10_58__Default_Changelist_.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_Checkout_at_2020-08-06_오후_10_58_[Default_Changelist]" date="1596722341174" recycled="false" toDelete="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_2020-08-06_오후_10_58_[Default_Changelist]/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 2020-08-06 오후 10:58 [Default Changelist]" />
+</changelist>
