+++ .idea/shelf/Uncommitted_changes_before_Checkout_at_2020-08-17__1_35__Default_Changelist_.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_Checkout_at_2020-08-17_오후_1_35_[Default_Changelist]" date="1597638942680" recycled="false" toDelete="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_2020-08-17_오후_1_35_[Default_Changelist]/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 2020-08-17 오후 1:35 [Default Changelist]" />
+</changelist>
