--- src/GeomKernel/__init__.py
@@ -1,1 +1,0 @@
-from .dataTypes import primitive as dtype
