--- src/GeomKernel/dataTypes/geometric/_GeomDataType.py
@@ -1,5 +1,0 @@
-from .._DataType import *
-
-
-class GeomDataType(ColumnVectorLikeData):
-    pass
