+++ src/gkernel/__init__.py
@@ -1,0 +1,1 @@
+from .dtype import primitive as dtype
