+++ src/gkernel/dtype/geometric/_GeomDataType.py
@@ -1,0 +1,5 @@
+from .._DataType import *
+
+
+class GeomDataType(VectorLikeData):
+    pass
