+++ src/gkernel/dtype/geometric/__init__.py
