+++ src/gkernel/dtype/geometric/primitive/__init__.py
@@ -1,0 +1,1 @@
+from .primitive import *
