+++ src/gkernel/dtype/nongeometric/_NoneGeomDataType.py
