+++ src/gkernel/dtype/nongeometric/matrix/__init__.py
